+++ oap-server/server-core/src/main/java/org/apache/skywalking/oap/server/core/analysis/worker/MetricsPersistentWorker.java
@@ -188,6 +188,10 @@ public class MetricsPersistentWorker extends PersistenceWorker<Metrics> implemen
      */
     @Override
     public void in(Metrics metrics) {
+        if (metricsDAO.isExpiredCache(model, metrics, System.currentTimeMillis(), metricsDataTTL)) {
+            log.warn("Receiving expired metrics: {}, time: {}, ignored", metrics.id(), metrics.getTimeBucket());
+            return;
+        }
         aggregationCounter.inc();
         dataCarrier.produce(metrics);
     }