+++ docs/en/changes/changes.md
@@ -81,6 +81,7 @@
 * Add more model installation log info for OAP storage initialization.
 * BanyanDB: Separate the storage configuration to an independent file: `bydb.yaml`.
 * Bump Armeria to 1.32.0 and some transitive dependencies.
+* Skip persisting metrics/record data that have been expired.
 
 #### UI